requirements.txt
Django==1.4
argparse==1.2.1
distribute==0.6.24
python-mpd2==0.4.2
wsgiref==0.1.2
